refactor estrutura
