+++ src/app/ui/components/navigation/Header.jsx
@@ -81,7 +81,7 @@ function Header() {
           <Link href="/login" title='Login'>
             <LogIn size={18} />
           </Link>
-          <Link href="/perfilt" title='Perfil'>
+          <Link href="/perfil" title='Perfil'>
             <UserCircle size={18} />
           </Link>
           <Search size={18} />
