+++ src/app/ui/components/partials/MovieWrapper.jsx
