+++ src/app/ui/components/sections/Novidades.jsx
@@ -1,9 +1,9 @@
 import { useContext } from "react";
 import { Swiper, SwiperSlide } from "swiper/react";
 import { Pagination, Navigation } from "swiper/modules";
-import { convertDate } from "../../api/utils/utils";
-import { DataContext } from "../../api/ContextApi";
-import Cartaz from "@/app/components/partials/Cartaz";
+import { convertDate } from "../../../data/utils/utils";
+import { DataContext } from "../../../data/hooks/ContextApi";
+import Cartaz from "@/app/ui/components/partials/Cartaz";
 import styles from "../../css/components/Novidades.module.scss";
 import "swiper/css";
 import "swiper/css/pagination";
