+++ src/app/ui/css/components/Sliders.module.scss
