+++ src/app/ui/css/pages/404.scss
@@ -33,7 +33,7 @@ body{
   height: 100%;
   top: 0;
   left: 0;
-  background: url("../../../../public/assets/filter-img.png");
+  background: url("../../../../../public/assets/filter-img.png");
   animation: interferenceStatic 3s steps(5, end) infinite;
   opacity: 0.65;
 }
