+++ src/app/ui/css/pages/Details.module.scss
