+++ src/app/ui/css/pages/Series.module.scss
